+++ pycuber/__init__.py
@@ -1,4 +1,4 @@
-from .cube import Cube, Face, Square, Step, Algo
+#from .cube import Cube, Face, Square, Step, Algo
 
 __title__ = "PyCuber"
 __version__ = "2.0.0"
